+++ CHANGELOG.md
@@ -5,17 +5,24 @@ npm run build
 npm version major/minor/patch
 npm publish --access=public
 
-npm install @mikezimm/fps-library-v2@1.0.80
+npm install @mikezimm/fps-library-v2@1.0.81
 
 # Changelog
 
 NOTE:  TURN BACK ON STRICT NULL CHECKS - turned off for PropPaneCols
 
+## 1.0.81 - 2023-Jan-03: ALVFinMan components
+- added makeBubbleElementFromBubbles, IMinTeachBubble to src/components/atoms/TeachBubble
+- refactored bubbles to include content element
+- Add FadeCarousel and SelectDots under src/components/molecules/FadeCarousel ( from test-fps-test/ALVFinMan web part )
+- relocated thinArray and getSpreadIndexes to src/logic/Arrays/thinOutArray.ts
+- created getWhoDunItElements in src/components/atoms/WhoDunItRow/getWhoDunIt.tsx to build Peope/Dates row
+
 ## 1.0.80 - 2022-Dec-31: clean up console.logs, ALVFinMan fetches
 - remove unneccessary console.logs
 - fix a couple check4Gulp imports that pull in pnp2
 - update PinMeGroup to only auto-expand on PageInfo for now
-- add getNextIndex function from FadeCarousel component to src\logic\Arrays\searching\getNextIndex.ts
+- add getNextIndex function from FadeCarousel component to src/logic/Arrays/searching/getNextIndex.ts
 - update getSourceItems to automatically create expandColumns from selectColumns if it is empty - ALVFinMan
 - add fetchFpsPageAsXML and fetchFpsItemAsXML to fetch Modern Page content